fikset mer validering

+++ src/webapp/controllers/AdminController.php
@@ -4,6 +4,7 @@ namespace tdt4237\webapp\controllers;
 
 use tdt4237\webapp\Auth;
 use tdt4237\webapp\models\User;
+use tdt4237\webapp\validation\UserNamePasswordValidation;
 
 class AdminController extends Controller
 {
@@ -34,17 +35,20 @@ class AdminController extends Controller
     }
 
     public function delete($username)
-    {
-        $isAdmin = $this->auth->user()->isAdmin();
-        if($isAdmin){
-            if ($this->userRepository->deleteByUsername($username) === 1) {
-                $this->app->flash('info', "Sucessfully deleted '$username'");
-                $this->app->redirect('/admin');
-                return;
+    {   
+        $validation = new UserNamePasswordValidation();
+        if($validation->validateUserName($username)) {
+            $isAdmin = $this->auth->user()->isAdmin();
+            if($isAdmin){
+                if ($this->userRepository->deleteByUsername($username) === 1) {
+                    $this->app->flash('info', "Sucessfully deleted '$username'");
+                    $this->app->redirect('/admin');
+                    return;
+                }
             }
-        }
-        else{
-           $this->app->redirect('/'); 
+            else{
+               $this->app->redirect('/'); 
+            }
         }
         
         $this->app->flash('info', "An error ocurred. Unable to delete user '$username'.");
@@ -53,18 +57,20 @@ class AdminController extends Controller
 
     public function deletePost($postId)
     {
-        $isAdmin = $this->auth->user()->isAdmin();
-        if($isAdmin){
-            if ($this->postRepository->deleteByPostid($postId) === 1) {
-                $this->app->flash('info', "Sucessfully deleted '$postId'");
-                $this->app->redirect('/admin');
-                return;
+        $validation = new UserNamePasswordValidation();
+        if($validation->validatePostId($postId)) {
+            $isAdmin = $this->auth->user()->isAdmin();
+            if($isAdmin){
+                if ($this->postRepository->deleteByPostid($postId) === 1) {
+                    $this->app->flash('info', "Sucessfully deleted '$postId'");
+                    $this->app->redirect('/admin');
+                    return;
+                }
+            }
+            else{
+              $this->app->redirect('/');  
             }
         }
-        else{
-          $this->app->redirect('/');  
-        }
-
         $this->app->flash('info', "An error ocurred. Unable to delete user '$username'.");
         $this->app->redirect('/admin');
     }
